--- a/src/com/projectkorra/ProjectKorra/ConfigManager.java
+++ b/src/com/projectkorra/ProjectKorra/ConfigManager.java
@@ -307,6 +307,7 @@ public class ConfigManager {
 				+ "Additionally, you can click while channeling to attack things near you, dealing damage and knocking them back. "
 				+ "Releasing shift at any time will dissipate the form.");
 		config.addDefault("Abilities.Water.OctopusForm.Range", 10);
+		config.addDefault("Abilities.Water.OctopusForm.AttackRange", 2.5);
 		config.addDefault("Abilities.Water.OctopusForm.Radius", 3);
 		config.addDefault("Abilities.Water.OctopusForm.Damage", 3);
 		config.addDefault("Abilities.Water.OctopusForm.Knockback", 1.75);