--- a/src/com/projectkorra/ProjectKorra/ConfigManager.java
+++ b/src/com/projectkorra/ProjectKorra/ConfigManager.java
@@ -77,6 +77,7 @@ public class ConfigManager {
 		config.addDefault("Properties.RegionProtection.RespectTowny", true);
 		config.addDefault("Properties.RegionProtection.RespectPreciousStones", true);
 		config.addDefault("Properties.RegionProtection.RespectLWC", true);
+		config.addDefault("Properties.RegionProtection.CacheBlockTime", 5000);
 
 		config.addDefault("Properties.TagAPI.Enabled", true);