--- a/src/com/projectkorra/ProjectKorra/ProjectKorra.java
+++ b/src/com/projectkorra/ProjectKorra/ProjectKorra.java
@@ -76,7 +76,7 @@ public class ProjectKorra extends JavaPlugin {
 		}
 
 		Methods.deserializeFile();
-
+		Methods.startCacheCleaner(Methods.CACHE_TIME);
 		new CraftingRecipes(this);
 	}