Major Performance Improvement:
    *IMPORTANT If you are a server owner please read*
- A new config option has been added called CacheBlockTime (default 5 seconds). If your server has specific regions that toggle bending off and on multiple times within the span of a few minutes, such as a custom ProBendingArena or Agni Kai Arena, you will probably want to keep the CacheBlockTime less than 3 seconds. If you disable bending in an area it will take up to CacheBlockTime to update.
    
- If your server does not have any of these areas that constantly deny or allow bending then you should increase the CacheBlockTime to 30 seconds or so.
    
- The Block Cache makes it so the server does not have to keep processing the same block multiple times in a row. For example, imagine you used PhaseChange, the block cache will log all of the ice blocks so that it does not have to keep checking if they are bendable blocks. Rather than processing the blocks 20 times a second, if you set the block cache to 30 seconds then each block will only be checked once every 30 seconds (600x less processing in this case).

Removed a leftover broadcast message:
- The plugin no longer broadcasts "RPG NOT DETECTED"
    
Fixed Lightning not arcing in water:
- After fixing the issue with Lightning passing through walls, a new bug arose causing it to not sub arc in water. Basically, the Lightning was hitting dirt in the water which was causing the entire ability to remove itself.
    
- Added a simple process cache to speed up performance while checking each individual particle. Previously, the ability was only checking a small portion of the lightning particles. Now it checks every particle to increase precision, consequently, it uses a cache to make sure we aren't wasting processing power.
    
OctopusForm - Added an AttackRange variable:
- OctopusForm range was only controlling the distance that a player could stream water from. The AttackRange configuration option changes the punch distance of OctopusForm.
    
AirBlast - Fixed a severe bug that was causing the plugin to crash:
    * Don't be this specific in the release notes, wouldn't want people to use this against unupdated plugins *
- If a player looked straight down and spammed shift/click then the AirBlast was continuing into the void and not deleting itself. It was then attempting to affect every entity that was currently spawned in the game, causing a huge amount of processing. When the plugin attempted to do a distance call on the AirBlast location and an entity location the result was Double.NaN (not a number). The plugin was then trying to set a velocity using NaN which was causing an instant crash. This bug was fixed by adding NaN checks at both the Location and Velocity.